LayoutTests/imported/w3c/web-platform-tests/trusted-types/tt-block-eval-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.


FAIL eval blocks if the default policy rejects a value. assert_throws_js: function "_ => {
      eval('a="hello there"');
    }" did not throw
